Completed the Lock function

+++ contracts/gov-locker/src/state.rs
@@ -1,5 +1,4 @@
-use cosmwasm_std::Timestamp;
-use cosmwasm_std::{Addr, Coin};
+use cosmwasm_std::{Addr, Coin, Decimal, Timestamp};
 use cw_storage_plus::{Item, Map};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -12,7 +11,7 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "snake_case")]
 pub struct PeriodWeight {
     pub period: u64,
-    pub weight: f64,
+    pub weight: Decimal,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
